Initial commit
This code is the WalkDir algorithm ripped out from MinIO's ListObject
logic. Changes had to be made to get this code to compile. I sadly
started versioning too late, otherwise it could have been even clearer
which parts of the code I had to change.

In most cases code should only be commented out and not removed entirely
to retain maximum resemblance to it's origin.