GnuPlot: Fit in range
Currently there's a spike in the runtime around 16,000 files. It makes
no sense to include this in the linear regression fit.

The fit is probably unnecessary anyway, but it's a good tool to have
under the belt and IMHO it helps further visualize how much the measured
points diverge from what we would expect.